+++ src/features/TutorSearch/ListElement.jsx
@@ -1,4 +1,4 @@
-function SortElement({ children, icon, isActive, onClick }) {
+function ListElement({ children, icon, isActive, onClick }) {
 	return (
 		<button
 			className={`flex items-center gap-2 text-lg py-1 px-8  w-full hover:border-mainPurple first-of-type:rounded-t-md last-of-type:rounded-b-md border-transparent border-2 border-whiteHover border-b-0 border-t-0 first-of-type:border-t-2 last-of-type:border-b-2 transition-colors duration-200`}
@@ -12,4 +12,4 @@ function SortElement({ children, icon, isActive, onClick }) {
 	);
 }
 
-export default SortElement;
+export default ListElement;
